7-1-26/FastAPI Mongo Atlas Vector/requirements.txt
fastapi
uvicorn
pymongo
requests
mongo
python-dotenv
https
